Fixed Win64 issues, added more tests...
git-svn-id: http://asmjit.googlecode.com/svn/trunk@237 79de4a12-e7ed-11dd-bce9-bf72c0239d3a

+++ AsmJit-1.0/AsmJit/AssemblerX86X64.cpp
@@ -1265,7 +1265,7 @@ void AssemblerCore::_emitInstruction(uint32_t code, const Operand* o0, const Ope
       {
         LabelData& l_data = _labelData[reinterpret_cast<const Label*>(o0)->getId() & OPERAND_ID_VALUE_MASK];
 
-        uint32_t hint = o1->isImm() ? reinterpret_cast<const Imm&>(*o1).getValue() : 0;
+        uint32_t hint = (uint32_t)(o1->isImm() ? reinterpret_cast<const Imm&>(*o1).getValue() : 0);
         bool isShortJump = false;
 
         // Emit jump hint if configured for that.
